pyproject.toml
[build-system]
requires = [
  "setuptools>=61.0",
]
build-backend = "setuptools.build_meta"

[project]
name = "fudge"
version = "0.0.1"
requires-python = ">=3.10"
dependencies = [
    "Django==4.2",
    "python-magic>=0.4.27",
]
